greatly advanced sync page, still untested

+++ m3u_library_manager/gui/workers/sync_workers.py
@@ -1,4 +1,5 @@
 # gui/workers/sync_workers.py
+
 from PyQt6.QtCore import QThread, pyqtSignal
 from pathlib import Path
 import asyncio
@@ -10,7 +11,7 @@ from core.sync.sync_operations import SyncOperations
 class ComparisonWorker(QThread):
     """Background worker for file comparison"""
     finished = pyqtSignal(ComparisonResult)
-    progress = pyqtSignal(float)
+    progress = pyqtSignal(int)  # Changed to int
     error = pyqtSignal(str)
     
     def __init__(self, comparator: FileComparator, playlist_path: Path,
@@ -30,7 +31,7 @@ class ComparisonWorker(QThread):
                     self.playlist_path,
                     self.local_base,
                     self.remote_base,
-                    lambda p: self.progress.emit(p)
+                    lambda p: self.progress.emit(int(p))  # Convert to int here
                 )
             )
             self.finished.emit(result)
@@ -40,7 +41,7 @@ class ComparisonWorker(QThread):
 class SyncWorker(QThread):
     """Background worker for sync operations"""
     finished = pyqtSignal()
-    progress = pyqtSignal(float)
+    progress = pyqtSignal(int)  # Changed to int
     error = pyqtSignal(str)
     
     def __init__(self, sync_ops: SyncOperations, playlist_path: Path,
@@ -65,7 +66,7 @@ class SyncWorker(QThread):
                     self.add_local,
                     self.remove_remote,
                     self.remove_local,
-                    lambda p: self.progress.emit(p)
+                    lambda p: self.progress.emit(int(p))  # Convert to int here
                 )
             )
             self.finished.emit()
